requirements.txt
requests
clickclick
stups-cli-support
